% Generated by authors
\docType{data}
\name{DRYAD_data}
\alias{DRYAD_data}
\title{Full gene expression data from original experiments}
\format{
\subsection{\code{DRYAD_data}}{

A data frame with 1027848 rows and 170 columns:
\describe{
\item{Cell_ID}{`character` vector - Unique Cell ID}
\item{Animal_ID}{`integer` vector - Unique Animal ID}
\item{Animal_sex}{`character` vector - Gender of Animal ("Male" or "Female")}
\item{Behavior}{`character` vector - Description of behavioural treatment ("Naive" means no treatment was given)}
\item{Bregma}{`double` vector - Approximate location of slice in "Bregma" coordinates}
\item{Centroid_X}{`double` vector - x-coordinate of the centroid position for the cell in µm}
\item{Centroid_Y}{`double` vector - y-coordinate of the centroid position for the cell in µm}
\item{Cell_class}{`character` vector - Major cell class assigned to cell ("Ambiguous" represents putative doublet cells and were not further analyzed)}
\item{Neuron_cluster_ID}{`character` vector - Neuronal cluster assigned to cell (Empty if cell is not a neuron)}
\item{gene_columns_135}{`double` vector - Expression value for a given gene of the next 135 columns measured in the combinatorial smFISH run were determined as the total counts per cell divided by the cell volume and scaled by 1000}
\item{gene_columns_21}{`double` vector - Expression value for a given gene of the next 21 columns measured in the non-combinatorial, sequential FISH rounds were arbitrary fluorescence units per µm^3, but the same scale for all cells. A value of "NaN" for Fos indicates that this gene was not measured in this animal. These are the last 21 columns in the data.}
\item{blanks_5}{`double` vector - Genes named "blank-" represents the measurements of barcodes not assigned to any RNA and which serve as blank controls. There are five blank controls.}
}
}
}
\source{
<https://datadryad.org/stash/dataset/doi:10.5061/dryad.8t8s248>
}
\usage{
DRYAD_data
}
\description{
A dataset of gene expression in mice from experiments by Moffit et al
}
\keyword{datasets}
