% Generated by authors
\docType{data}
\name{GEO_barcodes}
\alias{GEO_barcodes}
\title{Barcodes data for GEO data}
\format{
\subsection{\code{GEO_barcodes}}{

A data frame with 31299 rows and 1 columns:
\describe{
\item{barcode}{`character` vector - cell barcodes}
}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE113576>
}
\usage{
GEO_barcodes
}
\description{
A dataset of cell barcodes for the GEO data
}
\keyword{datasets}
