% Generated by authors
\docType{data}
\name{GEO_matrix}
\alias{GEO_matrix}
\title{mRNA count matrix for GEO data}
\format{
\subsection{\code{GEO_matrix}}{

A matrix of 84333635 observed non-zero values with 27998 rows (for genes) and 31299 columns (for cell samples):
\describe{
\item{i}{`integer` row indices of non-zero entries in 0-base - Genes}
\item{j}{`integer` column indices of non-zero entries in 0-base - Cell Samples}
\item{x}{`numeric` vector - the non-zero entry at position `(i, j)` - mRNA Counts}
\item{Dim}{`integer` of length 2 - the dimensions of the matrix - Number of Genes x Number of Cell Samples}
}
}
}
\source{
<https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE113576>
}
\usage{
GEO_matrix
}
\description{
A matrix of mRNA counts (gene expressions) across cell samples for the GEO data
}
\keyword{datasets}
