% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcmb.rate}
\alias{rcmb.rate}
\title{An example recombination rate dataset}
\format{A data frame with 478 rows and 6 variables
\describe{
  \item{CHR}{Chromosome}
  \item{RS}{SNP rs number}
  \item{POS}{Chromosomal position}
  \item{RATE}{recombination rate between the SNP and the SNP above}
}}
\usage{
rcmb.rate
}
\description{
This dataset is an example dataset of recombination rates between SNPs
The variables are as follows:
}
\keyword{datasets}

