% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp.all2}
\alias{snp.all2}
\title{SNPs in the PLINK files}
\format{A data frame with 10279 rows and 12 variables:
\describe{
  \item{ord}{Ordering of the SNPs}
  \item{RS}{SNP rs number}
  \item{CHR}{Chromosome}
  \item{POS}{Chromosomal position}
  \item{A1}{A1 allele}
  \item{A2}{A2 allele}
  \item{MAF}{Minor allele frequency}
  \item{NCHROBS}{Number of observed chromosomes used in MAF calculation}
}}
\usage{
snp.all2
}
\description{
A dataset containing the list of variables in the PLINK files
The variables are as follows:
}
\keyword{datasets}

